special_env_vars.py: add {FETCH,RESUME}COMMAND_SSH to the environ_filter set
Portage is supposed to prevent configuration variables in both the
FETCHCOMMAND_ and RESUMECOMMAND_ namespaces from being propagated to the
environment of the "ebuild.sh" utility and its subprocesses. To that
end, the "special_env_vars.py" unit defines the 'environ_filter' set,
which includes the following twelve variables among its members.

- FETCHCOMMAND
- FETCHCOMMAND_FTP
- FETCHCOMMAND_HTTP
- FETCHCOMMAND_HTTPS
- FETCHCOMMAND_RSYNC
- FETCHCOMMAND_SFTP
- RESUMECOMMAND
- RESUMECOMMAND_FTP
- RESUMECOMMAND_HTTP
- RESUMECOMMAND_HTTPS
- RESUMECOMMAND_RSYNC
- RESUMECOMMAND_SFTP

However, it does not account for the fact that the SSH protocol is
supported. Address this by incorporating both the 'FETCHCOMMAND_SSH' and
'RESUMECOMMAND_SSH' variables into the set. Assuming their default
values, this saves 568 bytes of environment space, as well as preventing
them from unnecessarily leaking into "environment.bz2" files.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>