writemsg: Use the proper base class TextIOBase instead of... ...StringIO for identifying all text I/O streams.
When passing a regular text file descriptor to writemsg(), some bug-avoidance
code was not identifying all of the file descriptor types that don't need
Unicode encoding. The file descriptor was being tested for being an instance of
io.StringIO instead of the base class for text streams (io.TextIOBase). As such,
passing a file descriptor that was opened in the most natural way for a text
file, for example

    fd = open(logfile, "w")
    writemsg("Hello, world!\n", fd=fd)

would cause an exception because Unicode encoding would be applied unnecessarily,
creating a binary object on which fd.write() would raise an exception.

Signed-off-by: John R. Graham <john_r_graham@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1462
Closes: https://github.com/gentoo/portage/pull/1462
Signed-off-by: Sam James <sam@gentoo.org>