save-ebuild-env.sh: reformat identifier lists, one per line
On the 15th July, the "save-ebuild-env.sh" unit was modified so as to
declare the lists of identifiers that require unsetting as arrays,
obviating the need for escaped <newline> characters as an awkward line
continuation device. Additionally, these lists were reordered with a
custom sorting algorithm, in an attempt to improve their legibility.

Reformat these lists once again, such that there is only ever one
identifier per line. I am now of the view that this is what I ought to
have done in the first place, for the reasons described herewith.

Firstly, humans are quite adept at reading columnated lists that are in
an alphabetical order.

Secondly, having there be only one identifier per line results in
unified diffs that are much easier for humans to fathom in the case that
identifiers are added, renamed and/or removed.

Thirdly, merge conflicts become less likely to occur. Indeed, I found
this change to be immediately beneficial in the course of rebasing my
patch queue.

Fourthly, contributors are duly relieved of the implied burden of
maintaining the layout of the lists, above and beyond their contents.

It should be noted that each group of identifiers has been sorted in a
manner that is equivalent to LC_ALL=C sort.

See-also: fd01f441afa9053cdbbf47252255d19693d299c9
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>