save-ebuild-env.sh: use MAPFILE to hold identifiers instead of REPLY
Presently, the __save_ebuild_env() function purposes the 'REPLY'
variable as an array for temporarily holding lists of identifiers that
are to be unset. The 'REPLY' variable was chosen on the basis that it
its value is normally controlled by the read builtin. That is, by using
it, the probability of a namespace conflict is reduced.

However, 'REPLY' is not normally an array variable. While this
observation is of no significance in and as of itself, some might find
it strange to see that particular variable be used as an array. In view
of this, purpose 'MAPFILE' as an array instead. For those that don't
know, that is the variable populated by the mapfile/readarray builtin,
provided that no name is specified.

See-also: fd01f441afa9053cdbbf47252255d19693d299c9
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>