process: do not poll join() in MultiprocessingProcesss if we can
Based on the initial analysis of Esteve Varela Colominas, the polling
on join() with the fixed 100ms delay incurs a significant performance
penalty, especially for short-lived processes. And since portage is
prone to spawning many of those, the penalty adds up easily.

Instead of pooling proc.join() with a fixed 100ms delay, we now use
the blocking variant of join() started in a thread which we await in
the coroutine. Unfortunately, we only can use this if the 'forkserver'
or 'spawn' start method is active for multiprocessing. Otherwise,
there is the risk of deadlocks due the interaction between threads and
the 'fork' start method.

Before this change:

( cd lib; python3.13 -m timeit 'import portage.process; portage.process.spawn("true")' )
2 loops, best of 5: 104 msec per loop

After this change:

( cd lib; python3.14 -m timeit 'import portage.process; portage.process.spawn("true")' )
1 loop, best of 5: 54.9 msec per loop

Note that the forkerserver in Python 3.14 seems to come with a
performance hit. While polling allowed spawn times ~5 ms, we can't seem
to get lower than 55 ms with forkserver.

Thanks to Zac Medico for input and assistance on this change.

Bug: https://bugs.gentoo.org/958635
Signed-off-by: Florian Schmaus <flow@gentoo.org>