phase-helpers.sh: refactor unpack() and have tar execute its decompressors
This commit subjects the unpack() function to an overall refactoring
while incorporating a fix for a regression that I was aware of and which
has since been independently identified and filed as bug #962874. The
changes are described herewith.

Declare all of the supported archive suffixes in a fashion that mirrors
the language of PMS. To wit, all documented suffixes are initially
declared as elements of the 'suffixes' array, case-sensitively. In turn,
that means that the following suffixes are now considered as first class
entities, as opposed to being supported as a consequence of the
machinations of the __unpack_tar() function.

- tar.bz
- tar.bz2
- tar.gz
- tar.lzma
- tar.Z

After having composed the 'suffixes' array, declare an associative array
by the name of 'suffix_by', where each key of the latter is an element
of the former. In the case that the effective EAPI is found not to be
case-sensitive, exploit the 'l' attribute of the declare/typeset
builtin, by which bash is able to induce a lowercase conversion upon
assignment. The latter array is prepared before the main loop begins and
makes it easy to determine whether a given suffix is recognised, whilst
doing away with the code pertaining to the now defunct 'suffix_known'
variable (which was not particularly digestible).

Declare and localise the 'XZ_OPT' variable in such a way that it is
rendered exportable. Do so before the main loop begins, duly
guaranteeing that the __makeopts_jobs() function shall be called just
once and obviating the need to incorporate the -T option of the xz(1)
utility at a later juncture.

Employ a regular expression to extract the suffix of each filename that
is considered. The pattern matches what is commonly referred to as an
extension whilst permitting for it to be preceded by "tar.". Once the
suffix has been extracted, employ a trivial -v test so as to determine
whether it exists as a key in the 'suffix_by' array.

Jettison the __unpack_tar() function. Not only was it responsible for
operating on files that are _not_ tarballs but it was also responsible
for further scrutinising the nature of the suffix. Long had I considered
it to be in poor taste. Instead, have each case pattern be accompanied
by the appropriate command.

To compensate for the aforementioned change, introduce the following
case patterns.

- tar.bz|tar.bz2|tbz|tbz2)
- tar|tar.*|tgz)

Further, jettison the following case patterns.

- tar)
- tbz|tbz2)
- tgz)

Jettison the __compose_bzip2_cmd() function. Instead, determine which
command may be employed to handle bzip2-compressed streams before the
main loop begins, unconditionally. Further, refrain from throwing an
exception in the event that both the 'PORTAGE_BUNZIP2_COMMAND' and
'PORTAGE_BZIP2_COMMAND' variables are defined as blank values.

Refrain entirely from employing shell pipelines to combine the tar(1)
utility with any particular decompression utility. Instead, delegate the
responsibility of executing the decompressor entirely to tar(1). The
rationale for this is twofold.

Firstly, GNU tar is already perfectly capable of recognising a
considerable number of file suffixes in the case that the -f option is
specified. No fewer than 16, in fact. Based on these suffixes, it is
further capable of determining whether to execute any of 8 distinct
decompression utilities: compress(1), bzip2(1), gzip(1), lzip(1),
lzma(1), lzop(1), xz(1) and zstd(1). Ergo, there is no sense in
maintaining equivalnt bash code that amounts merely to busywork.

Secondly, there have been several portage bugs concerning the use of
shell pipelines over the years. Some pertain to the matter of knowing
the exit status of the decompressor, while others pertain to the
consumption of malformed streams that cause tar(1) to close its end of
the pipe while the decompressor is still writing. By shifting the burden
of responsibility to tar(1), this entire class of problem is obviated.

It is worth noting that bzip2-compressed tarballs are handled as a
special case. The reason for this is straightforward: there is the
continued need to support the 'PORTAGE_BUNZIP2_COMMAND' and
'PORTAGE_BZIP2_COMMAND' portage variables. To accommodate these, the -I
option is used to convey a suitable decompression command. For instance,
to specify -I "lbzip2 -c" would imply to tar(1) that it must execute
lbzip2(1) with both the -c and -d options in effect. Though the -I
option-argument is not conveyed to a shell, it is parsed in a manner
that somewhat resembles that of a shell. Therefore, the values of the
aforementioned portage variables shall only be accepted if they consist
entirely of one or more non-blank characters that are neither the single
quote nor the double quote.

Fixes: 502bd866c85a44852bf1b214c3504d00ca50add5
Closes: https://bugs.gentoo.org/962874
Link: https://www.gnu.org/software/tar/manual/tar.html#auto_002dcompress
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>
(cherry picked from commit 3e89139fae34c9bd2e2b4c0490512f71d1d78546)
Signed-off-by: Sam James <sam@gentoo.org>