bintree: Accelerate index fetch by requesting Packages.gz first
Fetching the compressed package index (Packages.gz) significantly
reduces the time required to transfer the index. At the time of
writing this, the size of the uncompressed package index of the
official Gentoo binhost is ~16MiB, while the compressed variant's size
is ~1.7MiB.

Portage got support writing the a compressed package index with
11c0619c63b5 ("Portage writes a compressed copy of 'Packages' index
file."). However, this did not include *fetching* the compressed
package index, as the idea back then was to piggy-back on HTTP
compression. Unfortunately, many binhosts do not configure their HTTP
service accordingly.

This commit adjusts Portage logic to first try to fetch the compressed
package index with a fallback to the uncompressed one.

The diff of this change looks wild, but essentially it introduced only
a few new lines of code:
- a new loop: for remote_pkgindex_file in ("Packages.gz", "Packages"):
  - which required indenting the existing code by one level
- a break statement at the end of the loop (if it was successful)
- wrapping the existing file object into GzipFile if required via
  if remote_pkgindex_file == "Packages.gz":
      f = GzipFile(fileobj=f, mode="rb")
- a check if the received a 404 for Packages.gz, issuing the fallback
  to Packages in this case

Note that when ssh is used to fetch the file, we currently do not
request Packages.gz as it may not exist and the fallback is
non-trivial in this case.

Bug: https://bugs.gentoo.org/248089
Closes: https://bugs.gentoo.org/962719
Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1454
Closes: https://github.com/gentoo/portage/pull/1454
Signed-off-by: Sam James <sam@gentoo.org>