bintree: Suppress up-to-date message if it isn't true
Portage would emit a message claiming that the local copy of the remote
index is up-to-date when getbinpkg_refresh=False. However, if
getbinpkg_refresh is set to False, then this doesn't mean that the
local copy is up-to-date. It means that we didn't check if it's
up-to-date.

Since populate() is potentially multiple times, typically the second
time with getbinpkg_refresh set to False, this leads to multiple
"Local copy of remote index is up-to-date" messages, even if just one
binhost is configured.

To avoid this potentially confusing behavior, suppress the message in
case getbinpkg_refresh is set to False.

Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1456
Closes: https://github.com/gentoo/portage/pull/1456
Signed-off-by: Sam James <sam@gentoo.org>