bintree: Add missing 'break' in UseCachedCopyOfRemoteIndex
When we added support to fetch Packages.gz first, a new lopo was added
to the _populate_remote() logic, iterating over ("Packages.gz",
"Packages"). There are two situtations where we want to break out of
the loop: when the remote package index was fetched or when the cached
version of the package index was used. However,
fd55382a4935 ("bintree: Accelerate index fetch by requesting
Packages.gz first"), only added a break statement for the first
case (successful fetch). This adds the missing break statement for the
second case (use cached version).

Fixes: fd55382a4935 ("bintree: Accelerate index fetch by requesting Packages.gz first")
Signed-off-by: Florian Schmaus <flow@gentoo.org>