_urlopen: fix timestamp comparison by using UTC consistently
In bintree.py's _populate_remote, we call http_to_timestamp on the
timestamp header we get from the remote binhost. Unfortunately, that uses
mktime() which is documented to return *local time*:

> This is the inverse function of localtime(). Its argument is the struct_time or
> full 9-tuple (since the dst flag is needed; use -1 as the dst flag if it is unknown)
> which expresses the time in local time, not UTC.

Switch to using mktime_tz and parsedate_tz which operate in UTC instead. I've
changed timestamp_to_http as well just to be safe: I think this may fix
a potential issue with the value we send for If-Modified-Since, as
datetime.fromtimestamp() gives localtime.

This was manifesting as a warning with --getbinpkg:

> [gentoobinhost] WARNING: Service distfiles.gentoo.org did not respect If-Modified-Since.
>  Consider asking the service operator to enable support for If-Modified-Since or
>  using another service (local: 2025-09-12T12:43:46+01:00, remote: 2025-09-12T11:43:46+01:00).

... which would disappear when using TZ=UTC.

Bug: https://bugs.gentoo.org/962832
Signed-off-by: Sam James <sam@gentoo.org>