bintree: Improve UseCachedCopyOfRemoteIndex handling, report if TTL
Move the logic where parts of the user-exposed message is constructed
from the except block to right before we raise the
UseCachedCopyOfRemoteIndex exception. This means we can get rid of
some logic in the except block, which tried to determine what the
cause of the UseCachedCopyOfRemoteIndex exception was.

As benefit, we will no report if the use the cached copy if it is
within its TTL.

Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1469
Signed-off-by: Sam James <sam@gentoo.org>
(cherry picked from commit c83466a50efdee2cac81a5ed6a660c11b9ac746e)
Signed-off-by: Sam James <sam@gentoo.org>