repos.conf: set sync-git-verify-commit-signature to 'true'
https://wiki.gentoo.org/wiki/Portage_with_Git recommends

    eselect repository enable gentoo

to enable git-based sync of ::gentoo. However, since Portage does not
enable sync-git-verify-commit-signature per default, this ends up in
an insecure configuration (when compared with an rsync based sync).

Fix this by enabling the setting per default for the ::gentoo
repository in Portage.

Closes: https://bugs.gentoo.org/959831
Bug: https://bugs.gentoo.org/901385
Bug: https://github.com/gentoo/eselect-repository/issues/31
Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1474
Closes: https://github.com/gentoo/portage/pull/1474
Signed-off-by: Sam James <sam@gentoo.org>