phase-helpers.sh: install default docs where DOCS is undeclared, not unset
EAPI 6 introduced the einstalldocs() function. A recent commit altered
the behaviour of the function in such a way that, if the 'DOCS' variable
is found to be unset, a default set of documentation files shall be
installed. The prior behaviour was to do so in the event that the
variable is undeclared.

It transpires that there is a disagreement between the implementation
and the specification. Below is an excerpt of the pseudocode presented
by PMS, with some redactions made for the purpose of legibility.

 7: else if the DOCS variable is unset then
 8:   for all d matching the filename expansion of README* ChangeLog AUTHORS ...
 9:     if file d exists and has a size greater than zero then
10:       call dodoc with d as argument
11:     end if
12:   end for
13: end if

Note the use of the term, unset. In bash, it is quite possible to
declare a variable that remains in an unset state.

# ( DOCS=(); test -v DOCS || echo unset )
unset
# ( f() { local -a DOCS; test -v DOCS || echo unset; }; f )
unset

Effectively, PMS is in the wrong. By focusing unduly on the
specification, I inadvertently introduced a regression whereby, if
'DOCS' is set as an empty array, the function proceeds to install the
default documentation.

Address this issue by restoring the prior behaviour of checking whether
'DOCS' is undeclared. Further, perform the test by taking advantage of
the ${param@A} form of expansion, rather than parse the output of the
declare builtin as the prior implementation did. Not only does this
approach avoid an unnecessary subshell but it addresses an adjacent bug.
Imagine that 'DOCS' is declared as a function-local array that is empty.

src_install() {
	# This declares a local array that is unset and empty
	local -a DOCS
	# Imagine that einstalldocs() continues to parse declare -p.
	# Here, it will proceed to install the default documentation.
	einstalldocs
}

This long-standing bug was caused by testing in the fashion of
[[ $(declare -p DOCS) != *=* ]], which falsely assumes that the
<equals-sign> character is present for a declared variable.

# ( DOCS=(); declare -p DOCS; )
declare -a DOCS=()
# ( f() { local -a DOCS; declare -p DOCS; }; f ) # no <equals-sign>
declare -a DOCS

Fixes: 7ff579ccb92b2f2512933a41e9a0a204b0601201
Closes: https://bugs.gentoo.org/962934
Link: https://projects.gentoo.org/pms/6/pms.html#x1-145003r4
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>
(cherry picked from commit 5ee2d40e97c2aa21a3d322f486fe6783a150bbda)
Signed-off-by: Sam James <sam@gentoo.org>