emerge-webrsync: mark TZ as exportable in do_latest_snapshot()
As concerns the do_latest_snapshot() function, a recent commit modified
it so as to take advantage of the ability of the printf builtin to act
as a strftime(3) wrapper, rather than be beholden to the non-standard
date(1) utility. As such, it declares 'TZ' as a local variable, with the
intent of ensuring that the UTC timezone is in effect.

However, though printf is indeed a builtin, to assign a new value to the
'TZ' variable is insufficient, as is demonstrated beneath.

# cd /etc
# cp /usr/share/zoneinfo/America/New_York localtime
# TZ=UTC
# printf '%(%F)T\n' 1758758400
2025-09-24
# export TZ
# printf '%(%F)T\n' 1758758400
2025-09-25

Address this issue by setting the 'x' attribute for the 'TZ' variable at
the point of its declaration.

Reported-by: Hank Leininger <hlein@korelogic.com>
Closes: https://bugs.gentoo.org/963423
Fixes: caa8f296a5b3f5bdb8f36321a9334f8ff52d9d30
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>