phase-helpers.sh: unpack(): handle "--" in accordance with PMS
As concerns the unpack() function, a recent commit cleaned up its
option parsing routine. However, the revised code fails to account for
the fact that PMS mandates that the "--" argument be treated in a
somewhat unconventional way. Consider the following invocation.

  eapply -d "${S}/subdir" -- "${FILESDIR}/${PN}-some-necessary.patch"

Those familiar with getopt(3) conventions will immediately understand
that "--" is used to separate options and option-arguments from operands
(non-options). But the specification for eapply() goes further by
mandating that, if "--" is encountered, all arguments to its left are to
be conveyed to the patch(1) utility, irrespective of their values.
Further, in that event, the specification mandates that the rule of
options only being able to precede operands shall not be enforced.

Such an approach confers the benefit of being able to convey
option-arguments following short options to patch(1), without eapply()
needing to be aware of how the utility parses them. Therein lies the
rub, since the revised code fails to handle "--" in the prescribed way.
Instead, the "${S}/subdir" argument is considered as an eapply() operand
on account of not leading with a <hyphen-minus>.

Address this issue by once again permitting for the positional
parameters to be iterated over up to two times, and handling the
presence of the "--" argument as a special case. Yet, do so without
unduly impacting on the simplicity of the code that is a consequence of
its original refactoring.

Reported-by: Pavel Sobolev <contact@paveloom.dev>
Closes: https://bugs.gentoo.org/963536
Fixes: d594f7a29999779aa644c30a85a25c9811f02970
Link: https://projects.gentoo.org/pms/8/pms.html#x1-127001r1
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>