emerge-webrsync: restore the missing option-arg for curl --continue-at
The curl(1) user-agent supports a --continue-at option, which requires
an option-argument. Where the option-argument is <hyphen-dash>, curl
will automatically try to determine how to resume a transfer. Further,
emerge-webrsync forcibly injects the option if it detects that the value
of FETCHCOMMAND specifies a curl command.

However, the option-argument was unintentionally removed during a bout
of re-factoring and (otherwise unrelated) bug fixing. This went
unnoticed until now because wget is more commonly used. Restore it.

Fixes: f3af8263ce544750a9e9a155b1264d07a86677d4
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>