egencache: retry metadata process with unexpected returncode
Since MetadataRegen is not well suited for internal retry, create
a new MetadataRegen instance for each retry. Returns the returncode
from the last MetadataRegen instance, which will only be non-zero
if all retries failed.

Since we only observe bug 965132 in the python3.14 forkserver tests,
it seems like this must be some kind of forkserver bug that causes
our bash proccess to occasionally silently fail with returncode 255.

It may be very difficult to find the root cause of this forkserver
issue, and we can't predict how soon it will be fixed. Meanwhile,
I think we will appreciate the retry if it eliminates our
intermittent CI failures.

Another possible solution would be to force the multiprocessing
start method to spawn inside the egencache process (like
_EbuildFetcherProcess).

Bug: https://bugs.gentoo.org/965132
Signed-off-by: Zac Medico <zmedico@gentoo.org>