bintree: fix AttributeError by swapping try and for-loop
Swap the try-block and for-loop in _populate_remote().

Instead of the try-block being within the for-loop, the try-block is
no the outermost block containing the for-loop. This is sensible
because previously the beginning of the for-loop's body contained some
logic that was unrelated to the for-loop, i.e., the checks if the repo
is frozen or if the currently cached index is within its TTL.

This also fixes an AttributeError where in
    ttl = float(pkgindex.header.get("TTL", 0))
where pkgindex would be None in a subsequent iteration of the
loop. Now this code resides outside the for-loop (where it belongs).

This fixes the issue initially reported in

    https://bugs.gentoo.org/965098#c0

Furthermore, we move the finally block where we kill the download proc
and unlink the tmp file, which was previously attached to the
outermost try-block, to the innermost try-block because it needs to be
executed on every download attempt.

Bug: https://bugs.gentoo.org/965098
Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1490
Signed-off-by: Sam James <sam@gentoo.org>