bintree: Restore "except OSError" block
With the refactoring done in Fixes: 16655f8588e3 ("bintree: fix
AttributeError by swapping try and for-loop"), two "except OSError"
blocks were merged into one. This restores the outer "except OSError"
block, because it used to catch the OSError that was re-raised deeper
inside the functions logic. Otherwise, the re-raised OSError would
cause Portage to abort, which was previously not the case.

Fixes: 16655f8588e3 ("bintree: fix AttributeError by swapping try and for-loop")
Closes: https://bugs.gentoo.org/965464
Signed-off-by: Florian Schmaus <flow@gentoo.org>