ecompress: avoid losing track of find(1) in fix_symlinks()
Presently, the fix_symlinks() function is tasked with finding and
correcting symlinks that are dangling as a consequence of their targets
having been compressed. The routine that does so is structured in the
following way.

while IFS= read -rd '' link && IFS= read -rd '' target; do
	# Repair the symlinks. May execute rm(1) and ln(1).
done < <(
	# Find dangling symlinks.
	find "$top" -type l -xtype l -printf '%p\0%l\0'
)

# Reap the exit status of find(1), returning false upon failure.
wait "$!" || return

For this, a user reported the following error.

ecompress: line 172: wait: pid 18991 is not a child of this shell

After some initial difficulty in reproducing it, I realised that bash
was losing track of find(1) in its capacity as a background job, owing
to PID reuse. The reasons for the user being able to consistently induce
this issue are as follows.

- both rm(1) and ln(1) are executed many times for openssl (5138 each)
- his kernel was subject to a limit of kernel.pid_max = 32768
- he has a large number of processes (or endures a high fork rate)

Here is a simple reproducer.

# sysctl -w kernel.pid_max=512
# while read -r; do /bin/true; done < <(seq 1 1024); wait "$!"
bash: wait: pid 320 is not a child of this shell

Interestingly, not all shells have this problem. Consider dash as a case
in point.

# dash -c 'mkfifo fifo
{ seq 1 1024; exit 123; } > fifo & pid=$!
while read -r line; do /bin/true; done < fifo; wait "$pid"
echo "$?"'
123

I found it odd that the ability to reap a single asynchronous command
would be affected by proceeding to issue any number of commands in the
foreground. I contacted Chet Ramey, upon which he acknowledged the
problem and explained that bash is only able to keep one record of a
single PID throughout its job control code. Further, this limitation
extends to both background and foreground jobs, so as to simplify the
internal bookkeeping. Consequently, he is considering whether it would
be practical to have bash retain the status of a terminated background
process in the event that a foreground process recycles its PID.

In the meantime, ecompress must operate in a manner that is immune to
this issue. To that end, eschew the use of a process substitution by the
fix_symlinks() function. Instead, employ a conventional pipeline and
determine the exit status of find(1) by examining the PIPESTATUS array.
Also, to compensate for this change, enable the lastpipe shell option,
so that the assignment to the 'something_changed' variable is not lost
to a subshell.

Reported-by: Dan Johansson <dan.johansson@dmj.nu>
Fixes: 781732d87525469e311732d5418ffb0f3e419da8
Closes: https://bugs.gentoo.org/965423
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>