man: make.conf(5): mention that PORTAGE_COMPRESS="" inhibits compression
Some define PORTAGE_COMPRESS as "true" or ":" to inhibit compression but
the correct way to go about it is to define the compressor as the empty
string. Document the fact that it may be defined in this way.

Further, clarify that its value is treated as the name of a command.

See-also: 3498e0f1c1fe6c7f23707d0e56462fbc8dfb487d
Bug: https://bugs.gentoo.org/965903
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>