misc-functions.sh: don't run ecompress if PORTAGE_COMPRESS is empty
The 'PORTAGE_COMPRESS' variable allows for the user to select the
program used to compress documentation during the install phase. Those
wishing to disable compression may do so by defining the compressor as
an empty string. However, even in the case that compression is disabled,
the install_qa_check() function of the "misc-functions.sh" unit will
still execute the ecompress utility three times in succession.

  ecompress --queue "${PORTAGE_DOCOMPRESS[@]}"
  ecompress --ignore "${PORTAGE_DOCOMPRESS_SKIP[@]}"
  ecompress --dequeue

The outcome is that the 'queue' and 'ignore' phases are always executed
in full. Only the 'dequeue' phase recognises that compression has been
disabled, upon which it unlinks all of the files generated by the
'queue' phase before exiting. All of which is a waste of resources.

Address this issue by preventing the install_qa_check() function from
executing the ecompress utility in the case that the 'PORTAGE_COMPRESS'
variable is empty. Further, to compensate for this change, delegate the
responsibility of defining bzip2(1) as the default compressor to the
"make.globals" file.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>