file_copy: do not attempt to use _fastcopy on Darwin
The old linux reflink module would only trigger on Linux.  With the
introduction of util/file_copy it changed to always trying to exploit
faster ways of copying falling back to shutil.copyfile.  However, on
Darwin this always fails, and the fallback isn't silent (it generates
warnings and backtraces) so don't try on this platform to save time and
make the output nicer.

Bug: https://bugs.gentoo.org/966267
Signed-off-by: Fabian Groffen <grobian@gentoo.org>