binrepos.conf: implement `verify-signature`
Add `verify-signature` to binrepos.conf to allow controlling PGP signature
verification on a per-binary repository basis.

Users may want to verify binpkgs from the official Gentoo binhost or
other binhosts they run themselves, but not be forced to have a signing
key on such consumer machines to casually run `quickpkg` or `FEATURES=buildpkg`.

We can then toggle the default in the config file we ship in stages,
the handbook, etc, and possibly ship some default in Portage for `[gentoo]`. For
this work to be deployed, we also first need to change the default `location`
for that binhost to something else to avoid comingling of trusted and (potentially)
untrusted binpkgs.

I've not split this into two like we do for FEATURES="binpkg-request-signature"
and FEATURES="binpkg-ignore-signature". Instead, `verify-signature` just
controls 'both' (i.e. on = verify, off = don't even look to see if there's
a signature). I think having two toggles was a mistake we shouldn't repeat.

TODO: Make sure it's reliable to trust the `verify-signature` property from
      where we obtain it

TODO: Make things less noisy when we deliberately turn off verification
      and verification then fails (e.g. no keyring)

TODO: Tests (got a draft but it needs to know that failure for a case is OK,
      and need to add some other scenarios)

Bug: https://bugs.gentoo.org/930730
Bug: https://bugs.gentoo.org/941770
Bug: https://bugs.gentoo.org/945384
Signed-off-by: Sam James <sam@gentoo.org>