Implement FEATURES=packdebug
Create a tarball of debug information and source files for use with
binary packages and debuginfod.

The tarball is owned by the package, but we add it to UNINSTALL_IGNORE
so it is kept around on upgrades etc., and we add it to PKG_INSTALL_MASK
so that it is not part of binpkgs themselves (should be fetched via debuginfod).

That is, for FEATURES=packdebug, we:
* Create a tarball in /usr/lib/debug/.tarball/${CATEGORY}/${PN}/${PF}-${BUILD_ID}-debug.tar.xz
  with debug information from /usr/lib/debug/*, /usr/src/debug/*. This
  is installed but not part of the binpkg.
* Set UNINSTALL_IGNORE="/usr/lib/debug/.tarball" to keep old
  debug information tarballs around to serve.
* Set PKG_INSTALL_MASK="/usr/lib/debug/* /usr/src/debug/*" so binpkgs
  do not contain debug information (they should fetch it via debuginfod
  instead).

The model is as follows:
* binhost

  Builds packages with -g* in *FLAGS and has FEATURES=splitdebug
  (and possibly FEATURES=installsources) to get debuginfo. Must also
  have build IDs enabled in the compiler.

  Uses FEATURES=packdebug to create tarballs containing that info.

  Runs `debuginfod ... -Z tar.xz /usr/lib/debug/.tarball` to serve
  the tarballs generated by packdebug.

* binpkg client

  Sets the `DEBUGINFOD_URLS` environment variable to the httpd that
  `debuginfod` runs.

Thanks to Arsen for the original bashrc-based implementation which
the estrip impl. is based upon, motivation for getting this done, and
the idea. See also https://wiki.gentoo.org/wiki/User:Arsen/Deguginfod.

This is also really useful for not just serving binpkgs to clients and
letting them get debuginfo (of course the main usecase), but also getting
debuginfo for local binaries you don't even publish anywhere after an upgrade
when an old process is still running.

Note that xz is used for the debug tarball because it supports [0][1] random
access.

[0] https://blog.osandov.com/2024/07/25/making-debuginfod-viable-for-the-linux-kernel.html
[1] https://developers.redhat.com/articles/2025/01/14/debuginfod-project-update-2024#addressing_kernel_vdso_extraction_bottlenecks

Bug: https://bugs.gentoo.org/639376
Bug: https://bugs.gentoo.org/728818
Bug: https://bugs.gentoo.org/953869
Co-authored-by: Arsen Arsenović <arsen@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>