[WIP] jobserver support
Add FEATURES=jobserver-token that causes Portage to acquire a job token
from a jobserver (if one is running) for every started task, and release
it once it finishes.  Note that this does not implement the implicit
slot feature, since it is meant to work with the system-wide jobserver,
under the assumption that emerge will be run directly by the user
and therefore no job token will be taken for the emerge invocation
itself.

Closes: https://bugs.gentoo.org/966879
Signed-off-by: Michał Górny <mgorny@gentoo.org>