config: add getbinpkg-exclude and getbinpkg-include to binrepos.conf
Allow configuration in binrepos.conf of specific packages which are
to be satisfied (or not) using remote binary packages. These attributes
appear under the repository section and are specific to that binary
package host, with the behaviour otherwise identical to the command
line arguments of the same name.

Where atoms specified with getbinpkg-exclude or getbinpkg-include in
binrepos.conf match those for the opposing command line options then
emit a warning and override the binrepos.conf atoms.

Signed-off-by: Jethro Donaldson <devel@jro.nz>
Suggested-by: Sam James <sam@gentoo.org>