portage: add srcroot support to abssymlink()
Absolute symlinks inside an image directory currently resolve against the
host root, even though they are intended to be interpreted relative to the
image root (for example ${D}).

Add an optional srcroot argument so that absolute symlink dereferencing
behaves as if performed inside a chroot rooted at srcroot.

Historical behavior is preserved when srcroot is None.

Closes: https://bugs.gentoo.org/954361
Closes: https://bugs.gentoo.org/967661
Signed-off-by: Z. Liu <zhixu.liu@gmail.com>