lib/portage/package/ebuild/_config/LicenseManager.py: Read per-profile package.license
Currently the only way to unmask license in profile is to use the LICENSE
variable in make.defaults. This may be acceptable for "unique" licenses
(linux-fw-redistributable, only used by one package), but not practical for
"generic" licenses.

Read profile-specific package.license in LicenseManager so per-profile
per-package license can be configured.

Bug: https://bugs.gentoo.org/686888
Suggested-by: Sam James <sam@gentoo.org>
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Reviewed-by: Sam James <sam@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>