Scheduler: Only check free inode count on certain filesystems
The 'favail' value may be zero, or another made-up value, on
filesystems which allocate inodes dynamically (e.g., btrfs). Checking
for a certain number of free inodes is hence only sensible on
filesystems with a static inode count.

Thanks to dol-sen for reporting this.

Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1532
Closes: https://github.com/gentoo/portage/pull/1532
Signed-off-by: Sam James <sam@gentoo.org>